+++ gt_manage/miniprogram/pages/PageMan/refundOrder/refundOrder.js
@@ -1,20 +1,52 @@
 // miniprogram/pages/PageMan/refundOrder/refundOrder.js
+import { CloudFn } from '../../../utils/CloudFn'
+const cloudFn = new CloudFn()
 Page({
 
   /**
    * 页面的初始数据
    */
   data: {
-
+    refundArray: [], // 上架商品列表
   },
 
   /**
    * 生命周期函数--监听页面加载
    */
   onLoad: function (options) {
-
+    this._loadData()
   },
 
+
+  // 获取全部上架商品
+  _loadData: function() {
+    wx.showLoading({
+      title: '加载中...',
+      icon:'none'
+    })
+    cloudFn.$callFn({
+      data: {
+        name: 'CloudAPIBase',
+        body: {
+          fn: 'get',
+          base: 'user-order',
+          is_where: false,
+          where_data: {
+            state: 4,
+            refund_info: {
+              r_state: 1
+            }
+          }
+        }
+      }
+    }).then(res => {
+      wx.hideLoading()
+      this.data.refundArray = res.obj
+      this.setData({
+        refundArray: this.data.refundArray
+      })
+    })
+  },
   /**
    * 生命周期函数--监听页面初次渲染完成
    */