+++ changelog.md
@@ -1,4 +1,7 @@
 
+### 2020/09/16
+- gt_shop小程序 与 gt_manage小程序
+  - update user order module;
 
 ### 2020/09/15
 - gt_shop小程序